Test:  utLib1 Library CHECKX() changed to 64-bit (uint64_t).
    Test set t_utLib1/ also verifies ambiguity in CHECK() overload, which
    remains 32-bit.
    Makefile cleanup.

+++ librgpio/test/src/utLib1.h
@@ -51,8 +51,8 @@ class utLib1 {
 
     static void		ut_CHECKX(
 			    int                 line,
-			    unsigned int	ref,
-			    unsigned int	exp
+			    uint64_t		ref,
+			    uint64_t		exp
 			);
 };
